Rename single challenging protocol to dispute, add readme
