+++ packages/wallet/src/redux/protocols/container.tsx
@@ -1,7 +1,7 @@
 import { PureComponent } from 'react';
 import { ProtocolState } from '.';
 import * as fundingStates from './funding/states';
-import * as challengingStates from './challenging/state';
+import * as DisputeStates from './dispute/state';
 import * as concludingStates from './concluding/state';
 import React from 'react';
 import { Funding } from './funding/container';
@@ -9,7 +9,7 @@ import { Concluding } from './concluding/container';
 import { connect } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSpinner } from '@fortawesome/free-solid-svg-icons';
-import { Challenging } from './challenging/container';
+import { Challenging } from './dispute/container';
 
 interface Props {
   protocolState: ProtocolState;
@@ -23,7 +23,7 @@ class ProtocolContainer extends PureComponent<Props> {
     const { protocolState } = this.props;
     if (fundingStates.isNonTerminalFundingState(protocolState)) {
       return <Funding state={protocolState} />;
-    } else if (challengingStates.isNonTerminalChallengingState(protocolState)) {
+    } else if (DisputeStates.isNonTerminalDisputeState(protocolState)) {
       return <Challenging state={protocolState} />;
     } else if (concludingStates.isConcludingState(protocolState)) {
       return <Concluding state={protocolState} />;
