+++ packages/wallet/src/redux/protocols/dispute/challenger/components/approve-challenge.tsx
