+++ packages/wallet/src/redux/protocols/dispute/reducer.ts
@@ -1,7 +1,7 @@
 import { SharedData } from '../../state';
 import { Commitment } from '../../../domain';
 import { ProtocolStateWithSharedData } from '..';
-import { ChallengingState, isTerminal } from './state';
+import { DisputeState, isTerminal } from './state';
 import { initialize as responderInitialize, responderReducer } from './responder/reducer';
 import { initialize as challengerInitialize, challengerReducer } from './challenger/reducer';
 import { isNonTerminalResponderState } from './responder/state';
@@ -15,7 +15,7 @@ export const initializeResponderState = (
   channelId: string,
   sharedData: SharedData,
   challengeCommitment: Commitment,
-): ProtocolStateWithSharedData<ChallengingState> => {
+): ProtocolStateWithSharedData<DisputeState> => {
   return responderInitialize(processId, channelId, sharedData, challengeCommitment);
 };
 
@@ -23,7 +23,7 @@ export const initializeChallengerState = (
   processId: string,
   channelId: string,
   sharedData: SharedData,
-): ProtocolStateWithSharedData<ChallengingState> => {
+): ProtocolStateWithSharedData<DisputeState> => {
   const { storage, state: protocolState } = challengerInitialize(processId, channelId, sharedData);
   return { protocolState, sharedData: storage };
 };
@@ -32,7 +32,7 @@ export const challengingReducer = (
   protocolState: ChallengerState,
   sharedData: SharedData,
   action: ProtocolAction,
-): ProtocolStateWithSharedData<ChallengingState> => {
+): ProtocolStateWithSharedData<DisputeState> => {
   if (!isChallengerAction(action) && !isResponderAction(action)) {
     return { protocolState, sharedData };
   }
