+++ packages/wallet/src/redux/protocols/dispute/responder/components/wait-for-approval.tsx
